// frontend/vite-project/src/types/DataTypes.ts
export interface DisplayInterface {
    price: number;
    liveFlatPrice: number;
    imgSrc: string;
    liveBroadcastId: number;
    broadcastTitle: string;
}

export interface FirstDisplayInterface {
    price: number;
    liveBroadcastId: number;
    imgSrc: string;
    broadcastTitle: string;
    productName: string;
    liveFlatPrice: number;
    liveRatePrice: number;
}